requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
configparser==7.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.7
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.62.1
django-cors-headers==4.3.1
django-filter==24.2
django-jazzmin==3.0.0
django-modeltranslation==0.18.13
djangorestframework==3.15.1
drf-yasg==1.21.7
gunicorn==20.1.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
